LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/parsing/font-variant-serialization-expected.txt


FAIL checking serialized value of font-variant assert_equals: expected "tabular-nums" but got ""
